♻️ Make BlogPostCard UI component to be reusable
